app/cli/pyproject.toml
[project]
name = "bdsteam-cli"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = ["pydantic>=2.8", "pyyaml>=6"]

[project.scripts]
bdsteam = "cli.__main__:main"
